09-todo-list-migration/fastapi-app/pyproject.toml
[project]
name = "fastapi-todo-app"
version = "1.0.0"
description = "FastAPI TO-DO List API with async PostgreSQL and Redis"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "pydantic[email]>=2.0.0",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "aioredis>=2.0.1",
    "python-dotenv>=1.0.0",
    "mmh3>=4.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "httpx>=0.25.0",
    "pytest-asyncio>=0.21.0",
]

